move Attribute implementation to Python
